refactor: move usecases to a onwer folders
